renamed dataset and trainer modules
