+++ tests/models/test_query.py
@@ -25,7 +25,7 @@ from deeprankcore.domain.features import edgefeats as Efeat
 from deeprankcore.domain import targettypes as targets
 
 from deeprankcore.feature import sasa, atomic_contact, bsa, pssm, amino_acid
-from deeprankcore.DataSet import HDF5DataSet
+from deeprankcore.dataset import HDF5DataSet
 
 
 def _check_graph_makes_sense(g, node_feature_names, edge_feature_names):
